requirements.txt
streamlit
gspread
oauth2client
pandas
pyyaml
requests
beautifulsoup4
python-dateutil
